Refactor userSettingsProvider
