+++ src/components/forms/pool_actions/WithdrawForm/components/WithdrawPreviewModal/components/WithdrawSummary.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import useNumbers, { FNumFormats } from '@/composables/useNumbers';
-import useUserSettings from '@/composables/useUserSettings';
+import { useUserSettings } from '@/providers/user-settings.provider';
 import { Pool } from '@/services/pool/types';
 
 /**
