+++ src/providers/index.ts
@@ -4,6 +4,5 @@
  * If a provider is dependent on another it must be imported/exported below
  * the dependency.
  */
-export { default as UserSettingsProvider } from './user-settings.provider';
 export { default as TokenListProvider } from './token-lists.provider';
 export { default as TokensProvider } from './tokens.provider';
